microservices/startup-service/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
confluent-kafka==2.3.0

# OpenTelemetry
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-django==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-psycopg2==0.44b0
opentelemetry-exporter-otlp==1.23.0
